+++ src/types/transcript.ts
@@ -7,17 +7,17 @@ export type SpeakerInfo = {
 
 export type Speaker = {
   speaker: string;
-  speaker_info?: SpeakerInfo;
+  speakerInfo?: SpeakerInfo;
   text?: string;
   words?: string[];
-  start_times?: number[];
+  startTimes?: number[];
 };
 
 export type Transcript = {
-  event?: EarningsEvent;
-  text?: string;
+  event: EarningsEvent;
+  text: string;
   speakers?: Speaker[];
-  prepared_remarks?: string;
-  questions_and_answers?: string;
-  speaker_name_map_v2?: { [key: string]: SpeakerInfo };
+  preparedRemarks?: string;
+  questionsAndAnswers?: string;
+  speakerNameMapV2?: { [key: string]: SpeakerInfo };
 };