+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.0.24](https://github.com/EarningsCall/earningscall-js/compare/v0.0.23...v0.0.24) (2024-12-13)
+
+-   BREAKING CHANGE: Convert types from snake case to camel case.
+-   Add `Symbols` class to help with symbol lookups.
+
 ## [0.0.23](https://github.com/EarningsCall/earningscall-js/compare/v0.0.22...v0.0.23) (2024-12-12)
 
 -   Try yet again to resolve type build issue.